[FSTORE-1476] Deduplicate connection (#286)
* Move connection

* Deduplicate connections

* Rework deduplication using lazy imports

* Make hsfs.engine self-initializing

* Moce usage to hopsworks_common

* Add usage alias to hsfs

* Fix docgen

* Fix mocking

* Fix connection so that hsml works

Many things in hsml expect load_default_configuration of ModelServingApi to be called in connect; it sets up different global variables of client (e.g., _serving_resource_limits).

* Fix connection, regarding the two-phase connect

* Fix mistype

* Remove redundant usage.py from hsml

It was not there before.
