+++ python/hsfs/engine/__init__.py
@@ -17,28 +17,25 @@ from __future__ import annotations
 
 from typing import TypeVar, Union
 
+import hopsworks_common.connection
 from hsfs.client import exceptions
 from hsfs.core import arrow_flight_client
 from hsfs.engine import spark, spark_no_metastore
 
 
 _engine = None
-_engine_type = None
 
 
 def init(engine_type: str) -> None:
-    global _engine_type
     global _engine
     if not _engine:
         if engine_type == "spark":
-            _engine_type = "spark"
             _engine = spark.Engine()
         elif engine_type == "hive":
             raise ValueError(
                 "Hive engine is not supported in hopsworks client version >= 4.0."
             )
         elif engine_type == "spark-no-metastore":
-            _engine_type = "spark-no-metastore"
             _engine = spark_no_metastore.Engine()
         elif engine_type in ["python", "training"]:
             try:
@@ -49,24 +46,16 @@ def init(engine_type: str) -> None:
                     "missing in HSFS installation. Install with `pip install "
                     "hsfs[python]`."
                 ) from err
-            _engine_type = "python"
             _engine = python.Engine()
-        elif engine_type == "training":
-            _engine = "training"
 
 
 def get_instance() -> (
     Union[spark.Engine, spark_no_metastore.Engine, TypeVar("python.Engine")]
 ):
     global _engine
-    if _engine:
-        if _engine == "training":
-            raise Exception(
-                "`training` engine doesn't support this operation. "
-                "Supported engines are `'spark'` and `'python'`."
-            )
-        return _engine
-    raise Exception("Couldn't find execution engine. Try reconnecting to Hopsworks.")
+    if not _engine:
+        init(hopsworks_common.connection._hsfs_engine_type)
+    return _engine
 
 
 # Used for testing
@@ -74,16 +63,14 @@ def set_instance(
     engine_type: str,
     engine: Union[spark.Engine, spark_no_metastore.Engine, TypeVar("python.Engine")],
 ) -> None:
-    global _engine_type
     global _engine
-    _engine_type = engine_type
+    hopsworks_common.connection._hsfs_engine_type = engine_type
     _engine = engine
 
 
 def get_type() -> str:
-    global _engine_type
-    if _engine_type:
-        return _engine_type
+    if _engine:
+        return hopsworks_common.connection._hsfs_engine_type
     raise Exception("Couldn't find execution engine. Try reconnecting to Hopsworks.")
