+++ python/tests/engine/test_python.py
@@ -23,7 +23,6 @@ import polars as pl
 import pyarrow as pa
 import pytest
 from hsfs import (
-    engine,
     feature,
     feature_group,
     feature_view,
@@ -43,7 +42,7 @@ from hsfs.training_dataset_feature import TrainingDatasetFeature
 from polars.testing import assert_frame_equal as polars_assert_frame_equal
 
 
-engine._engine_type = "python"
+hopsworks_common.connection._hsfs_engine_type = "python"
 
 
 class TestPython:
@@ -2554,7 +2553,7 @@ class TestPython:
     def test_apply_transformation_function_pandas(self, mocker):
         # Arrange
         mocker.patch("hopsworks_common.client.get_instance")
-        engine._engine_type = "python"
+        hopsworks_common.connection._hsfs_engine_type = "python"
         python_engine = python.Engine()
 
         @udf(int)
@@ -2594,7 +2593,7 @@ class TestPython:
     def test_apply_transformation_function_multiple_output(self, mocker):
         # Arrange
         mocker.patch("hopsworks_common.client.get_instance")
-        engine._engine_type = "python"
+        hopsworks_common.connection._hsfs_engine_type = "python"
         python_engine = python.Engine()
 
         @udf([int, int], drop=["col1"])
@@ -2638,7 +2637,7 @@ class TestPython:
         # Arrange
         mocker.patch("hopsworks_common.client.get_instance")
 
-        engine._engine_type = "python"
+        hopsworks_common.connection._hsfs_engine_type = "python"
         python_engine = python.Engine()
 
         @udf([int, int])
@@ -2689,7 +2688,7 @@ class TestPython:
         # Arrange
         mocker.patch("hopsworks_common.client.get_instance")
 
-        engine._engine_type = "python"
+        hopsworks_common.connection._hsfs_engine_type = "python"
         python_engine = python.Engine()
 
         @udf([int, int], drop=["col1", "col2"])
@@ -2735,7 +2734,7 @@ class TestPython:
         # Arrange
         mocker.patch("hopsworks_common.client.get_instance")
 
-        engine._engine_type = "python"
+        hopsworks_common.connection._hsfs_engine_type = "python"
         python_engine = python.Engine()
 
         @udf([int, int], drop=["col1"])
@@ -2783,7 +2782,7 @@ class TestPython:
         # Arrange
         mocker.patch("hopsworks_common.client.get_instance")
 
-        engine._engine_type = "python"
+        hopsworks_common.connection._hsfs_engine_type = "python"
         python_engine = python.Engine()
 
         @udf(int)
